Update update-immich.sh
enter version for updating manually

+++ update-immich.sh
@@ -1,8 +1,21 @@
-cd /tmp
-rm -rf immich-native
-git clone https://github.com/arter97/immich-native
-cd immich-native
-./install.sh
+#!/bin/bash
 
-systemctl restart immich
-systemctl status immich
+if [[ $# -eq 1 ]]; then
+  version="$1"
+
+  echo "Upgrading to Version $version";
+
+  cd /tmp
+  rm -rf immich-native
+  git clone https://github.com/arter97/immich-native
+  cd immich-native
+
+  sed -i "s/^REV=v.*/REV=v$version/" "install.sh"
+
+  ./install.sh
+
+  systemctl restart immich
+  systemctl status immich
+else
+  echo "enter version (X.Y.Z) as argument"
+fi
